Code - Adding two new functions - CvPlayerAI::AI_magicCombatValue(UnitTypes eUnit) and CvUnit::isRangedCollateral()
First pass at this. Kind of in beta test at the moment.

AI_magicCombatValue() tries to return a numeric value for a unit, defining how much combat value should be added due to spells and abilities

isRangedCollateral() returns true for a unit if they have a spell or ability that allows them to damage units at range