Fix - Item duplication bug
After picking up the item, it could still linger as a delayed death unit. Then,
it could be captured by other players, overriding the delayed death status.
I think the reason the delayed death units are not cleaned up in time is due to
the Sephi AI casting spells in CvGame::updateMoves().
It is not clear to me why the pick up item spells use delayed death, though.

--HG--
branch : mnai-fixes